separate coco_eval out and transforms
